Block List: Extract copy logic to separate non-visual component
